Configured Guano parsing to be less strict so it doesn't choke when Loc Accuracy is a float. Updated the external guano code to expect Loc Accuracy to be a float, per the Guano spec.

+++ batogram/external/guano.py
@@ -167,7 +167,7 @@ class GuanoFile(object):
 
     _coersion_rules = {
         'Filter HP': float, 'Length': float, 'Loc Elevation': float,
-        'Loc Accuracy': int, 'Samplerate': int,
+        'Loc Accuracy': float, 'Samplerate': int,
         'TE': lambda value: int(value) if value else 1,
         'Loc Position': lambda value: tuple(float(v) for v in value.split()),
         'Timestamp': parse_timestamp,
